Lots of little code changes
This commit shouldn't change the logic.

This is just changes to Clients

Big thanks to jetbrains for giving me a
license to resharper.  Most of these changes
come from resharper suggestions.

+++ src/Clients/MainApp/FSpot.Imaging/IOChannel.cs
@@ -210,7 +210,7 @@ namespace FSpot.Imaging {
 			if (error != IntPtr.Zero)
 				throw new GException (error);
 
-			return (int)read;
+			return read;
 		}
 
 		[DllImport("libglib-2.0-0.dll")]
@@ -261,7 +261,7 @@ namespace FSpot.Imaging {
 
 		public override long Seek (long position, SeekOrigin origin)
 		{
-      		      throw new NotSupportedException ();
+				  throw new NotSupportedException ();
 		}
 
 		[DllImport("libglib-2.0-0.dll")]
